client: deduplicate produceBlockWith4844Tx
